Minor fix
git-svn-id: https://svn.aspect.dealii.org/trunk/aspect@1064 244a8438-9d6d-46c2-a098-b74c898f6b90

+++ source/simulator/parameters.cc
@@ -236,7 +236,7 @@ namespace aspect
                          "initial global refinement but while still within the first "
                          "time step.");
       prm.declare_entry ("Time steps between mesh refinement", "10",
-                         Patterns::Integer (1),
+                         Patterns::Integer (0),
                          "The number of time steps after which the mesh is to be "
                          "adapted again based on computed error indicators. If 0 "
                          "then the mesh will not be refined.");
